New structure for all js created
