+++ Stylesheets/projects.css
@@ -110,7 +110,7 @@ main{
 /*-------------------------------------------------------------*/
 
 #img_card_donut{
-    background: url(../Images/Carousel_Donut.png);
+    background: url(../Images/carousel_donut.png);
     background-size:cover;
 }
