+++ index.html
@@ -18,7 +18,7 @@
     <div id="container">
         <header>
             <div class="flexbox" id="logoDiv">
-                <a href="#" id="logo-container"><img id="logo" src="./Images/JanLehner_Logo.png"
+                <a href="#" id="logo-container"><img id="logo" src="./Images/jan_lehner_logo.png"
                         alt="Picture of my logo"></a>
             </div>
         </header>
